WIP: refactor - MarkdownListUtil
